Split frontend script
