+++ package.json
@@ -5,8 +5,8 @@
 	"author": "Rory Heaney",
 	"license": "GPL-2.0-or-later",
 	"scripts": {
-		"build": "wp-scripts build",
-		"start": "wp-scripts start",
+		"build": "wp-scripts build src/index.js src/frontend.js",
+		"start": "wp-scripts start src/index.js src/frontend.js",
 		"format": "wp-scripts format",
 		"lint:css": "wp-scripts lint-style",
 		"lint:js": "wp-scripts lint-js",
