+++ src/frontend.js
@@ -1,0 +1,3 @@
+import lazyLoadVideos from './lazyVideos';
+
+lazyLoadVideos();
