+++ src/index.js
@@ -14,108 +14,7 @@ import './formats/span-format.js';
 // Import SCSS for compilation
 import './editor.scss';
 
-// WordPress dependencies
-import { registerBlockExtension } from '@10up/block-components';
+// Import helper modules
+import registerExtensions from './registerExtensions';
 
-// Import components
-import BlockEdit from './components/BlockEdit';
-
-// Import utilities
-import { generateClassName, generateAttributes } from './utils/helpers';
-
-// Import configurations
-import { ALLOWED_BLOCKS, BLOCK_CONFIG } from './config/blockConfig';
-
-// Register extensions for each block
-ALLOWED_BLOCKS.forEach( ( blockName ) => {
-	const config = BLOCK_CONFIG[ blockName ] || {};
-	const dropdownConfig = config.dropdown || {};
-
-	// Define attributes
-	const attributes = {
-		...generateAttributes(), // Dynamically generate padding, margin, and negative margin attributes
-	};
-
-	// Add FormTokenField attributes
-	for ( const classType of config.classOptions || [] ) {
-		attributes[ `${ classType }Classes` ] = {
-			type: 'array',
-			items: { type: 'string' },
-			default: [],
-		};
-	}
-
-	// Add dropdown attribute if applicable
-	if ( dropdownConfig.attributeKey ) {
-		attributes[ dropdownConfig.attributeKey ] = {
-			type: 'string',
-			default: dropdownConfig.default || 'none',
-		};
-	}
-
-	// Add width attributes for core/column
-	if ( config.hasWidthControls ) {
-		attributes.widthBase = { type: 'string', default: '' };
-		attributes.widthSm = { type: 'string', default: '' };
-		attributes.widthMd = { type: 'string', default: '' };
-		attributes.widthLg = { type: 'string', default: '' };
-		attributes.widthXl = { type: 'string', default: '' };
-		attributes.widthXXl = { type: 'string', default: '' };
-	}
-
-	// Register the extension
-	registerBlockExtension( blockName, {
-		extensionName: `custom-${ blockName.replace( 'core/', '' ) }`,
-		attributes,
-		Edit: BlockEdit,
-		classNameGenerator: ( attrs ) =>
-			generateClassName( attrs, blockName, BLOCK_CONFIG ),
-	} );
-} );
-
-// Lazy load cover background videos
-document.addEventListener( 'DOMContentLoaded', () => {
-	// .wp-block-cover__video-background[data-src]
-	//  video[data-fs-lazy-video][data-src]
-	const lazyVideos = document.querySelectorAll(
-		'video[data-fs-lazy-video][data-src]'
-	);
-
-	// console.log( 'Lazy loading videos:', lazyVideos );
-
-	// If no lazy videos, exit early
-	if ( lazyVideos.length === 0 ) {
-		return;
-	}
-
-	if ( 'IntersectionObserver' in window ) {
-		const options = {
-			root: null,
-			rootMargin: '200px',
-			threshold: 0,
-		};
-		const observer = new window.IntersectionObserver( ( entries ) => {
-			entries.forEach( ( entry ) => {
-				if ( entry.isIntersecting ) {
-					const video = entry.target;
-					const dataSrc = video.getAttribute( 'data-src' );
-					if ( dataSrc ) {
-						video.setAttribute( 'src', dataSrc );
-						video.removeAttribute( 'data-src' );
-					}
-					observer.unobserve( video );
-				}
-			} );
-		}, options );
-
-		lazyVideos.forEach( ( video ) => observer.observe( video ) );
-	} else {
-		lazyVideos.forEach( ( video ) => {
-			const dataSrc = video.getAttribute( 'data-src' );
-			if ( dataSrc ) {
-				video.setAttribute( 'src', dataSrc );
-				video.removeAttribute( 'data-src' );
-			}
-		} );
-	}
-} );
+registerExtensions();
